backend/weather-service/requirements.txt
pandas
numpy
kafka-python
pytest
python-dotenv
shapely 
requests 
geopy